Refactor time formatting logic: move formatTimeUntil to a new module and update imports
